Add the sqlite3_result_error_code() application interface.  Use it in the ATTACH function so that a failed attach returns a proper error code. Ticket #2914. (CVS 4775)
FossilOrigin-Name: c24616204307936d03d39d2ef0fe6856113f6977

+++ src/vdbeInt.h
@@ -209,7 +209,7 @@ struct sqlite3_context {
   VdbeFunc *pVdbeFunc;  /* Auxilary data, if created. */
   Mem s;                /* The return value is stored here */
   Mem *pMem;            /* Memory cell used to store aggregate context */
-  u8 isError;           /* Set to true for an error */
+  int isError;          /* Error code returned by the function. */
   CollSeq *pColl;       /* Collating sequence */
 };
